Domains Update
1. Separate domain index.ts into individual ts files for simplicity
2. Use identity for domain and update refs
