+++ v0.4/platform/domains/src/model_storage.ts
@@ -1,6 +1,21 @@
+//
+// Copyright © 2020, 2021 Anticrm Platform Contributors.
+//
+// Licensed under the Eclipse Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License. You may
+// obtain a copy of the License at https://www.eclipse.org/legal/epl-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
 import { Class, Doc, DocumentQuery, FindOptions, Model, Ref, Storage, Tx, TxContext } from '@anticrm/core'
-import { CORE_CLASS_CREATE_TX, CORE_CLASS_DELETE_TX, CORE_CLASS_UPDATE_TX, CreateTx, DeleteTx, UpdateTx } from '.'
-import { updateDocument } from './tx/modeltx'
+import domains, { CreateTx, DeleteTx, UpdateTx } from '.'
+import { updateDocument } from './tx'
 
 export class ModelStorage implements Storage {
   model: Model
@@ -25,15 +40,15 @@ export class ModelStorage implements Storage {
 
   async tx (ctx: TxContext, tx: Tx): Promise<any> {
     switch (tx._class) {
-      case CORE_CLASS_CREATE_TX: {
+      case domains.class.CreateTx: {
         const createTx = tx as CreateTx
         return this.model.add(this.model.createDocument(createTx._objectClass, createTx.object, createTx._objectId))
       }
-      case CORE_CLASS_UPDATE_TX: {
+      case domains.class.UpdateTx: {
         const updateTx = tx as UpdateTx
         return updateDocument(this.model, this.model.get(updateTx._objectId), updateTx.operations)
       }
-      case CORE_CLASS_DELETE_TX: {
+      case domains.class.DeleteTx: {
         const deleteTx = tx as DeleteTx
         return this.model.del(deleteTx._objectId)
       }
