+++ v0.4/platform/domains/src/tx/modeltx.ts
@@ -1,5 +1,20 @@
+//
+// Copyright © 2020, 2021 Anticrm Platform Contributors.
+//
+// Licensed under the Eclipse Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License. You may
+// obtain a copy of the License at https://www.eclipse.org/legal/epl-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
 import { AnyLayout, ArrayOf, AttributeMatch, Class, CORE_CLASS_ARRAY_OF, CORE_CLASS_INSTANCE_OF, CORE_CLASS_OBJ, Model, Obj, Ref } from '@anticrm/core'
-import { ObjectSelector, TxOperation, TxOperationKind } from '..'
+import { ObjectSelector, TxOperation, TxOperationKind } from '../tx'
 /**
  * Perform update of document attributes
  * @param doc - document to update
